Remove dependency on Rails
This commit removes the development dependency on Rails. This required
several changes to the code.

First, we switched to Minitest without using ActiveSupport. The test
setup code is now placed in test/setup.rb and is responsible for loading
all dependencies and preparing the test environment. We no longer split
the code into several support files.

Second, production code uses explicit requires in order not to rely on
eager loading.

Third, the dummy test app was removed as it's no longer needed.

Unfortunately, we had to remove the test case ensuring all tasks are
reported by rake -T as it's no longer possible to implement without
Rails.