Add support for project-wide configuration files
We add support for per-project configuration for active_record_doctor
detectors and initialization. It's a sequence of related changes that
we'll discribe separately.

active_record_doctor/config is responsible for loading and processing of
configuration files. It contains a loader class used to parse the config
and one public API method intended for use from within the config.
Additionally, it performs extensive checks to ensure configuration is
valid.

Another important addition is active_record_doctor/errors. It's
currently used mostly by config but is generic and can be used by
other parts of the code base. It's primary goal is providing adequate
and detailed error feedback to users so that they can resolve problems
quickly and efficiently.

Since the config is now responsible for initialization, specifically
loading models to be inspected, we had to add a Runner class tying
configuration and detectors together. The initialization changes allowed
us to remove Rails::TestApplication from the test setup code.
Eager-loading is also removed from individual detectors.

The last big part is a Rake tasklib for defining all Rake tasks at once.
As a result, Rake-specific code is now located in a specialized class
that can be used in non-Rails Rakefiles as well as the .rake file that
ships with active_record_doctor.

Last but not least, we had to make some adjustments to the test suite
to make it work with the new code.