Fix missing_unique_indexes on smaller indexes
missing_unique_indexes used to expect an index on scope + column but
this is too restrictive. _Any_ unique index on a sublist of columns is
enough to guarantee uniqueness at the database level. This commit
lifts that restriction and also skips partial indexes as they cannot
be used to guarantee global uniqueness.