Fix incorrect_dependent_option on through associations
This commit fixes incorrect_dependent_option on through associations.
The problem was that in those cases the detector would look at the
final instead of the join model. Additionally, if the join model lacks
the corresponding has_many association a NoMethodError would be raised.
This case is now detected and signaled in a graceful manner.

Last but not least, the error message was fixed to mention "callbacks"
instead of "validations".