missing_non_null_contraint: support belongs_to_required_validates_foreign_key
This commit adds awareness for a new active record setting that could lead to
skipping validations that should be otherwise checked. Basically, the validator
skips conditional validations as if presence is validate conditionally then it
means it can be absent some of the time, so it makes no sense to recommend NOT
NULL.

Newer versions of Active Record implemented an optimization where the presence
of an association is validated only if the foreign key changed (otherwise we
know the record is still there thanks to database constraints). We can detect
this is a built-in validation and recommend it's backed by NOT NULL even though
it's conditional.