Add Ruby 3.4 support (#225)
Starting from Ruby 3.4, mutex_m has been changed from a default gem to a bundled gem. In systems using Bundler (such as Rails applications), the gem needs to be explicitly listed in the Gemfile and installed. This change ensures that the mutex_m library is properly available in Ruby 3.4 environments.