Rewrite missing_presence_validation to handle polymorphic associations (#230)
This commits rewrites missing_presence_validation in a more imperative
style with extensive commenting to make it clear why certain actions are
performed. The new structure made it simple to correctly handle
polymorphic associations: if a polymorphic foreign key is NOT NULL then
the detector will suggest making the association as required, instead of
suggesting adding validators for the _id and _type column separately.